freelance_ai/requirements.txt
httpx>=0.27,<0.28
beautifulsoup4>=4.12,<5.0
SQLAlchemy>=2.0,<2.1
python-telegram-bot>=20.8,<21.0
APScheduler>=3.10,<4.0
python-dotenv>=1.0,<2.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
